requirements.txt
streamlit
faiss-cpu
langchain
langchain-community
beautifulsoup4
requests
openai
pickle-mixin
tiktoken